Move math methods to calls_math.py (#418)
Moves 18 methods to the new CallsMathMixin:
- Math: abs, min, max, floor, ceil, round, sqrt, pow
- Numeric conversions: to_float, float_to_int, nat_to_int,
  int_to_nat, byte_to_int, int_to_byte
- Float predicates: is_nan, is_infinite, nan, infinity

calls.py: 8,390 -> 7,948 lines (-442)
calls_math.py: 0 -> 457 lines

Also adds calls_math and the other new calls_* modules to the
mypy mixin override list in pyproject.toml.
