+++ pyproject.toml
@@ -103,6 +103,14 @@ module = [
     "vera.wasm.inference",
     "vera.wasm.operators",
     "vera.wasm.calls",
+    "vera.wasm.calls_arrays",
+    "vera.wasm.calls_containers",
+    "vera.wasm.calls_encoding",
+    "vera.wasm.calls_handlers",
+    "vera.wasm.calls_markup",
+    "vera.wasm.calls_math",
+    "vera.wasm.calls_parsing",
+    "vera.wasm.calls_strings",
     "vera.wasm.closures",
     "vera.wasm.data",
 ]
